+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
 ### 0.1.2
 
-* Removing println! and replacing them with proper error handling and error propagation
+* If the configuration file cannot be read, use the default values
+* Removing some logs and replacing them with proper error handling and error propagation
 * Added Connection resiliency. If the RabbitMq dies, it will start retrying to connect until it exhausts all retries
 
 ### 0.1.1